+++ oxe-api/resource/private/update_my_user.py
@@ -29,6 +29,7 @@ class UpdateMyUser(MethodResource, Resource):
         'telephone': fields.Str(required=False, allow_none=True),
         'accept_communication': fields.Bool(required=False),
         'accept_request_notification': fields.Bool(required=False),
+        'accept_terms_and_conditions': fields.Bool(required=False),
         'vcard': fields.Str(required=False, allow_none=True),
         'is_vcard_public': fields.Bool(required=False),
     })