+++ oxe-web-admin/src/component/PageProfile.jsx
@@ -382,10 +382,16 @@ export default class PageProfile extends React.Component {
 					<div className="col-md-8">
 						<div className={"row row-spaced"}>
 							<div className="col-md-12 PageProfile-white-box">
-								<h3>Administration</h3>
+								<h3>Notification</h3>
 								<br/>
 								<FormLine
-									label={"Accept request notification"}
+									label={"Accept to receive communications"}
+									type={"checkbox"}
+									value={this.state.user.accept_communication}
+									onChange={(v) => this.updateUser("accept_communication", v)}
+								/>
+								<FormLine
+									label={"Accept request notifications (for admins only)"}
 									type={"checkbox"}
 									value={this.state.user.accept_request_notification}
 									onChange={(v) => this.updateUser("accept_request_notification", v)}