+++ oxe-web-admin/src/component/item/user/UserGlobal.jsx
@@ -100,13 +100,19 @@ export default class UserGlobal extends React.Component {
 						/>
 						<br/>
 						<FormLine
+							label="Accept terms and conditions"
+							type={"checkbox"}
+							value={this.state.user.accept_terms_and_conditions}
+							disabled={true}
+						/>
+						<FormLine
 							label="Accept to receive communications"
 							type={"checkbox"}
 							value={this.state.user.accept_communication}
 							disabled={true}
 						/>
 						<FormLine
-							label="Accept to receive request notifications (only for active admins)"
+							label="Accept request notifications (for admins only)"
 							type={"checkbox"}
 							value={this.state.user.accept_request_notification}
 							disabled={true}